move opensea files
